nvshmem: fix pd allocation for dct object
DCT related code will be removed eventually.
For now just make logical fix for completeness.

PD for DCT object is not device level PD.
It is as separate PD, so create new PD.

Note - PD for DCT QP in current NVSHMEM implementation use
Domain level PD which is not supported and it is commented
in code as TBD.

Bug: DCSG-34135
Change-Id: Iaf7a022eb8654ddee9df4bc072b5028d541247f4
Signed-off-by: Kashyap Desai <kashyap.desai@broadcom.com>
Reviewed-on: https://gerrit-ccxsw.broadcom.net/c/brcm_nvshmem/+/232283
Reviewed-by: Selvin Thyparampil Xavier <selvin.xavier@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ajit Khaparde <ajit.khaparde@broadcom.com>